Add isShipDirectionGood() and more functionality
Added a parameter to addShip() which allows the class to now place an entire ship instead of only one space. isShipDirectionGood() is directly tied to this, it checks to see if there is enough space to place the ship without it running into another ship or going off the grid.